修复mem-bw不设带宽限制时错误的输出warning信息的行为 整理rbench.cpp关于limit部分的常数到一个static map 新增core-net相关代码 修改help信息

+++ rbench-mem-bw.cpp
@@ -118,7 +118,7 @@ static void mem_bw_bench_module( int32_t thrid , bench_args_t args , uint64_t ar
                 args.mem_bandwidth + MEMBW_CONTROL_RBOUND < actl_membw ){
                 membw_to_time( BYTES_ADD_KERNEL , args.mem_bandwidth , sgl_time , sgl_idle , args.period , module_runrounds , module_sleepus ) ;
             }
-            if( args.mem_bandwidth - 100 * MB > actl_membw ){
+            if( (int64_t)( args.mem_bandwidth - 100 * MB ) > actl_membw ){
                 if( ++low_actl_membw_warning > 8 ){
                     sprintf( infobuf , "LOW MEM_BW - %s( thread %d ): current %.1fMB/s, target %.1fMB/s, adjusting..." , 
                         args.bench_name.c_str() , thrid , actl_membw / MB , (double)args.mem_bandwidth / MB ) ;
